+++ src/interfaces/items/IInfoPanelItem.ts
@@ -142,10 +142,17 @@ export interface IInfoPanelItem {
   body: IBox;
 
   /**
+   * The property that controls whether the header is visible in the info panel.
+   * By default, the header is visible.
+   *
+   */
+  isHeaderVisible?: boolean;
+
+  /**
    * A function that is executed after opening a tab.
    * It returns a new body. If this functionality is not needed, the old body value is returned.
    */
-  onLoad: () => Promise<{ body: IBox }>;
+  onLoad?: () => Promise<{ body: IBox }>;
 
   /**
    * The types of files where the current item will be displayed in the info panel.